Surplus Saver/requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
redis 
celery
python-jose[cryptography]